requirements.txt
PyMySQL==0.9.2
seaborn==0.9.0
cvxopt==1.2.2
SQLAlchemy==1.2.1
pandas>=0.23.4
requests==2.21.0
urllib3==1.22
numpy>=1.16.3
scipy==1.2.0
beautifulsoup4==4.8.2
python_dateutil==2.8.1
PyYAML==5.3
